Fixes
Fixed warnings and errors when compiling with different kernel versions

+++ ndpi-netfilter/src/ndpi_proc_info.c
@@ -225,11 +225,11 @@ ssize_t nproto_proc_read(struct file *file, char __user *buf,
 		if(!n->mark[i].mark && !n->mark[i].mask)
 		    l += snprintf(&lbuf[l],sizeof(lbuf)-l,"%02x  %17s %-16s # %lld \n",
 				i,"disabled",c_buf,
-				atomic64_read(&n->protocols_cnt[i]));
+				(long long int)atomic64_read(&n->protocols_cnt[i]));
 		else
 		    l += snprintf(&lbuf[l],sizeof(lbuf)-l,"%02x  %8x/%08x %-16s # %lld debug=%d \n",
 				i,n->mark[i].mark,n->mark[i].mask,c_buf,
-				atomic64_read(&n->protocols_cnt[i]),
+				(long long int)atomic64_read(&n->protocols_cnt[i]),
 #ifdef NDPI_ENABLE_DEBUG_MESSAGES
 					n->debug_level[i]
 #else
